--- a/files/usr/share/cinnamon/applets/grouped-window-list@cinnamon.org/applet.js
+++ b/files/usr/share/cinnamon/applets/grouped-window-list@cinnamon.org/applet.js
@@ -131,9 +131,6 @@ class PinnedFavs {
             opts.app = appSystem.lookup_app(opts.appId);
         }
         if (!opts.app) {
-            opts.app = appSystem.lookup_settings_app(opts.appId);
-        }
-        if (!opts.app) {
             opts.app = appSystem.lookup_desktop_wmclass(opts.appId);
         }
         if (!opts.app) {
@@ -663,9 +660,6 @@ class GroupedWindowListApplet extends Applet.Applet {
             for (let i = 0, len = specialApps.length; i < len; i++) {
                 if (specialApps[i].wmClass === wmClass) {
                     app = this.appSystem.lookup_app(specialApps[i].id);
-                    if (!app) {
-                        app = this.appSystem.lookup_settings_app(specialApps[i].id);
-                    }
                     if (app) {
                         app.wmClass = wmClass;
                     }