--- a/files/usr/share/cinnamon/applets/menu@cinnamon.org/applet.js
+++ b/files/usr/share/cinnamon/applets/menu@cinnamon.org/applet.js
@@ -3258,7 +3258,7 @@ class CinnamonMenuApplet extends Applet.TextIconApplet {
                 for (let i in this._applicationsButtons) {
                     let app = this._applicationsButtons[i].app;
                     if (Util.latinise(app.get_name().toLowerCase()).indexOf(pattern)!==-1 ||
-                        (app.get_keywords() && Util.latinise(app.get_keywords().toLowerCase()).indexOf(pattern)!==-1) ||
+                        (app.get_keywords() && Util.latinise(app.get_keywords().join(" ").toLowerCase()).indexOf(pattern)!==-1) ||
                         (app.get_description() && Util.latinise(app.get_description().toLowerCase()).indexOf(pattern)!==-1) ||
                         (app.get_id() && Util.latinise(app.get_id().slice(0, -8).toLowerCase()).indexOf(pattern)!==-1))
                         res.push(app.get_id());