--- a/src/cinnamon-app-private.h
+++ b/src/cinnamon-app-private.h
@@ -22,11 +22,6 @@ void _cinnamon_app_add_window (CinnamonApp *app, MetaWindow *window);
 
 void _cinnamon_app_remove_window (CinnamonApp *app, MetaWindow *window);
 
-void _cinnamon_app_do_match (CinnamonApp         *app,
-                          GSList           *terms,
-                          GSList          **prefix_results,
-                          GSList          **substring_results);
-
 G_END_DECLS
 
 #endif /* __CINNAMON_APP_PRIVATE_H__ */