--- a/src/cinnamon-app-system.c
+++ b/src/cinnamon-app-system.c
@@ -675,6 +675,7 @@ _cinnamon_app_system_notify_app_state_changed (CinnamonAppSystem *self,
   switch (state)
     {
     case CINNAMON_APP_STATE_RUNNING:
+    case CINNAMON_APP_STATE_BUSY:
       g_hash_table_insert (self->priv->running_apps, g_object_ref (app), NULL);
       break;
     case CINNAMON_APP_STATE_STARTING: