--- a/src/cinnamon-app-system.c
+++ b/src/cinnamon-app-system.c
@@ -40,7 +40,7 @@ struct _CinnamonAppSystemPrivate {
 
   GHashTable *running_apps;
   GHashTable *id_to_app;
-  GHashTable *startup_wm_class_to_app;
+  GHashTable *startup_wm_class_to_id;
 
   GSList *known_vendor_prefixes;
 
@@ -118,9 +118,8 @@ cinnamon_app_system_init (CinnamonAppSystem *self)
                                            NULL,
                                            (GDestroyNotify)g_object_unref);
 
-  priv->startup_wm_class_to_app = g_hash_table_new_full (g_str_hash, g_str_equal,
-                                                         NULL,
-                                                         (GDestroyNotify)g_object_unref);
+  priv->startup_wm_class_to_id = g_hash_table_new (g_str_hash, g_str_equal);
+
 
 /* According to desktop spec, since our menu file is called 'cinnamon-applications', our
  * merged menu folders need to be called 'cinnamon-applications-merged'.  We'll setup the folder
@@ -150,7 +149,7 @@ cinnamon_app_system_finalize (GObject *object)
   g_hash_table_destroy (priv->running_apps);
   g_hash_table_destroy (priv->id_to_app);
   g_hash_table_destroy (priv->setting_id_to_app);
-  g_hash_table_destroy (priv->startup_wm_class_to_app);
+  g_hash_table_destroy (priv->startup_wm_class_to_id);
   g_slist_free_full (priv->known_vendor_prefixes, g_free);
   priv->known_vendor_prefixes = NULL;
 
@@ -390,7 +389,7 @@ on_apps_tree_changed_cb (GMenuTree *tree,
           old_startup_wm_class = g_desktop_app_info_get_startup_wm_class (old_info);
 
           if (old_startup_wm_class)
-            g_hash_table_remove (self->priv->startup_wm_class_to_app, old_startup_wm_class);
+            g_hash_table_remove (self->priv->startup_wm_class_to_id, old_startup_wm_class);
 
           _cinnamon_app_set_app_info (app, info);
           g_object_ref (app);  /* Extra ref, removed in _replace below */
@@ -406,8 +405,8 @@ on_apps_tree_changed_cb (GMenuTree *tree,
 
       startup_wm_class = g_desktop_app_info_get_startup_wm_class (info);
       if (startup_wm_class)
-        g_hash_table_replace (self->priv->startup_wm_class_to_app,
-                              (char*)startup_wm_class, g_object_ref (app));
+        g_hash_table_replace (self->priv->startup_wm_class_to_id,
+                              (char*)startup_wm_class, (char*)id);
 
     }
   /* Now iterate over the apps again; we need to unreference any apps
@@ -619,10 +618,16 @@ CinnamonApp *
 cinnamon_app_system_lookup_startup_wmclass (CinnamonAppSystem *system,
                                             const char     *wmclass)
 {
+  const char *id;
+
   if (wmclass == NULL)
     return NULL;
 
-  return g_hash_table_lookup (system->priv->startup_wm_class_to_app, wmclass);
+  id = g_hash_table_lookup (system->priv->startup_wm_class_to_id, wmclass);
+  if (id == NULL)
+    return NULL;
+
+  return shell_app_system_lookup_app (system, id);
 }
 
 /**