--- a/src/cinnamon-app-system.c
+++ b/src/cinnamon-app-system.c
@@ -208,7 +208,7 @@ cinnamon_app_system_lookup_app (CinnamonAppSystem   *self,
     return NULL;
 
   app = _cinnamon_app_new (info);
-  g_hash_table_insert (priv->id_to_app, shell_app_get_id (id), app);
+  g_hash_table_insert (priv->id_to_app, (char *) shell_app_get_id (id), app);
   return app;
 }