This PR backports commits made to appsys upstream.

In particular, it removes the dependency on cinnamon-menus and makes AppSys standalone.

Note though that some of the components in Cinnamon still use gir1.2-cmenus:

- cinnamon menu editor
- menu applets (core and spices)

Appsys is embedded in Cinnamon and does not support categories, so it can't replace cmenus in its current form.