@clefebvre 2019-01-14
Backported up to:

https://github.com/GNOME/gnome-shell/commits/master/src/shell-app-system.c
May 28th 2014.

https://github.com/GNOME/gnome-shell/commits/master/src/shell-app.c
Nov 14, 2013