@clefebvre 2019-01-15
https://github.com/GNOME/gnome-shell/commits/master/src/shell-app-system.c
Nov 17th 2018.