@clefebvre 2019-01-22
Performance looks largely unchanged without HDD cache:

- master: 57MB RAM, 4.1s startup (1.6s appsys, 700ms menu)
- PR: 58MB RAM, 4.0s startup (1.3s appsys, 950ms menu)

With HDD cache, we're seeing a slightly faster init:

- master: 52ms appsys + 500 menu
- PR: 33ms appsys + 510 menu

TLDR: Performance isn't much different.