@clefebvre 2019-01-22
The main pros of this PR are:

- It updates some really old code (some elements in appsys go all the way back to 2012)
- It makes appsys similar to upstream (that will ease future backports)
- It completely separates cmenu from appsys, and that makes it simpler for us to optimize the menu applet.