@clefebvre 2019-01-22
The following functions are removed:

```
appsys.lookup_setting()
appsys.lookup_settings_app()
appsys.lookup_settings_app_by_tree_entry()
appsys.search_settings()

```
These functions return NULL since Cinnamon 1.8 (Mint 14), so we can remove usage safely.

Also removed are the unused:

```
appsys.lookup_app_for_path()
appsys.initial_search()
appsys.subsearch()
app.get_faded_icon()
```

Also removed is:

`apps.sys.get_tree()`

which can be replaced by (that code doesn't need the changes in this PR): 

```
tree = new CMenu.Tree({ menu_basename: "cinnamon-applications.menu" });
tree.load_sync();
```

And:

`appsys.lookup_app_by_tree_entry(entry)`

replaced with:

`appsys.lookup_app(entry.get_desktop_file_id())`

^^ this can be changed in applets without this PR.

Finally, the following changes are brought by this PR:

`app.get_keywords()` no longer returns a string, but an array of stings.. so it should be replaced by `app.get_keywords().join(" ")`.

appInfo.should_show() should be tested to know if an app should be visible:

`if (appInfo && !appInfo.get_nodisplay()) ` --> `if (appInfo && appInfo.should_show() && !appInfo.get_nodisplay())`