@clefebvre 2019-01-22
Almost everything is ready on master and spices for this to be merged.

app.get_keywords() is used in Cinnamenu@json. @jaszhix what do you think? Do you want to make its code compatible with both old and new (i.e. string and string array), or start a 4.2 version for it?

appinfo.get_nodisplay() is used in:

- Cinnamenu@json
- tracker@glebihan
- CinnVIIStarkMenu@NikoKrause
- officeCenter@scollins
- graphicsCenter@scollins

@glebihan @NikoKrause @collinss @jaszhix, this isn't critical but it might be a good idea to review whether or not appInfo.should_show() should be used in combination with appInfo.get_nodisplay(), not just for this PR, but for current Cinnamon/spices.