@jaszhix 2019-01-22
@clefebvre I would be open to starting a 4.2 version for it.

After looking at the glib source, it appears they both do the same thing. `g_desktop_app_info_get_nodisplay`is a little simpler and might perform better.