@clefebvre 2019-01-22
Ah cool, I'll try to get this merged tomorrow.