@collinss 2019-01-23
@clefebvre I'm not sure what should_show() does differently than get_nodisplay() (without looking at the code), but it seems to me that get_nodisplay() already gives me everything I need there, so I don't think I'll make any additional changes. I also refactored Office Center and Graphics Center the other day so that they don't use appsys anymore, and rely solely on CMenu, so I don't have a problem with this being merged on my end.